Add SMSG_FORCE_MOVE_(UN)ROOT and SMSG_STANDSTATE_UPDATE for wrath
