+++ wow_message_parser/wowm/world/movement/smsg/smsg_force_move_root.wowm
@@ -1,7 +1,13 @@
-#tag_all versions "1.12";
-
 smsg SMSG_FORCE_MOVE_ROOT = 0x00E8 {
     Guid guid;
     u32 counter;
+} {
+    versions = "1.12";
 }
 
+smsg SMSG_FORCE_MOVE_ROOT = 0x00E8 {
+    PackedGuid guid;
+    u32 counter;
+} {
+    versions = "3.3.5";
+}
