+++ wow_message_parser/wowm/world/movement/smsg/smsg_force_move_unroot.wowm
@@ -1,7 +1,14 @@
-#tag_all versions "1.12";
-
 smsg SMSG_FORCE_MOVE_UNROOT = 0x00EA {
     Guid guid;
     u32 counter;
+} {
+    versions = "1.12";
+}
+
+smsg SMSG_FORCE_MOVE_UNROOT = 0x00EA {
+    PackedGuid guid;
+    u32 counter;
+} {
+    versions = "3.3.5";
 }
