+++ wow_message_parser/wowm/world/social/smsg_standstate_update.wowm
@@ -1,4 +1,4 @@
-#tag_all versions "1.12";
+#tag_all versions "1.12 3.3.5";
 
 smsg SMSG_STANDSTATE_UPDATE = 0x029D {
     UnitStandState state;
