+++ wow_world_messages/src/world/shared/mod.rs
@@ -99,9 +99,11 @@ pub mod smsg_pong123;
 pub mod smsg_procresist1_1223;
 pub mod smsg_realm_split2_4_33;
 pub mod smsg_server_message1_122;
+pub mod smsg_standstate_update1_123_3_5;
 pub mod smsg_time_sync_req2_4_33;
 pub mod smsg_tutorial_flags123;
 pub mod transport_info1_122;
+pub mod unit_stand_state1_123_3_5;
 pub mod update_flag1_122_4_3;
 pub mod update_type1_122_4_3;
 pub mod vector3d123;
