+++ wow_world_messages/src/world/wrath/mod.rs
@@ -428,6 +428,10 @@ pub(crate) mod smsg_environmentaldamagelog;
 pub use smsg_environmentaldamagelog::*;
 pub(crate) mod smsg_feature_system_status;
 pub use smsg_feature_system_status::*;
+pub(crate) mod smsg_force_move_root;
+pub use smsg_force_move_root::*;
+pub(crate) mod smsg_force_move_unroot;
+pub use smsg_force_move_unroot::*;
 pub(crate) mod smsg_initialize_factions;
 pub use smsg_initialize_factions::*;
 pub(crate) mod smsg_login_settimespeed;
@@ -446,6 +450,8 @@ pub(crate) mod smsg_realm_split;
 pub use smsg_realm_split::*;
 pub(crate) mod smsg_server_message;
 pub use smsg_server_message::*;
+pub(crate) mod smsg_standstate_update;
+pub use smsg_standstate_update::*;
 pub(crate) mod smsg_text_emote;
 pub use smsg_text_emote::*;
 pub(crate) mod smsg_time_sync_req;
@@ -468,6 +474,8 @@ pub(crate) mod spline_flag;
 pub use spline_flag::*;
 pub(crate) mod transport_info;
 pub use transport_info::*;
+pub(crate) mod unit_stand_state;
+pub use unit_stand_state::*;
 pub(crate) mod update_flag;
 pub use update_flag::*;
 pub(crate) mod update_type;
