+++ wow_world_messages/src/world/wrath/smsg_standstate_update.rs
@@ -1,0 +1,2 @@
+pub use crate::shared::smsg_standstate_update1_123_3_5::SMSG_STANDSTATE_UPDATE;
+
