+++ wow_world_messages/src/world/wrath/unit_stand_state.rs
@@ -1,0 +1,2 @@
+pub use crate::shared::unit_stand_state1_123_3_5::UnitStandState;
+
