+++ wowm_language/src/docs/smsg_force_move_unroot.md
@@ -4,7 +4,7 @@
 
 ### Wowm Representation
 
-Autogenerated from `wowm` file at [`wow_message_parser/wowm/world/movement/smsg/smsg_force_move_unroot.wowm:3`](https://github.com/gtker/wow_messages/tree/main/wow_message_parser/wowm/world/movement/smsg/smsg_force_move_unroot.wowm#L3).
+Autogenerated from `wowm` file at [`wow_message_parser/wowm/world/movement/smsg/smsg_force_move_unroot.wowm:1`](https://github.com/gtker/wow_messages/tree/main/wow_message_parser/wowm/world/movement/smsg/smsg_force_move_unroot.wowm#L1).
 ```rust,ignore
 smsg SMSG_FORCE_MOVE_UNROOT = 0x00EA {
     Guid guid;
@@ -28,4 +28,35 @@ SMSG have a header of 4 bytes.
 | ------ | ----------------- | ---- | ---- | ----------- | ------- |
 | 0x04 | 8 / Little | [Guid](../spec/packed-guid.md) | guid |  |  |
 | 0x0C | 4 / Little | u32 | counter |  |  |
+
+# SMSG_FORCE_MOVE_UNROOT
+
+## Client Version 3.3.5
+
+### Wowm Representation
+
+Autogenerated from `wowm` file at [`wow_message_parser/wowm/world/movement/smsg/smsg_force_move_unroot.wowm:8`](https://github.com/gtker/wow_messages/tree/main/wow_message_parser/wowm/world/movement/smsg/smsg_force_move_unroot.wowm#L8).
+```rust,ignore
+smsg SMSG_FORCE_MOVE_UNROOT = 0x00EA {
+    PackedGuid guid;
+    u32 counter;
+}
+```
+### Header
+
+SMSG have a header of 4 bytes.
+
+#### SMSG Header
+
+| Offset | Size / Endianness | Type   | Name   | Description |
+| ------ | ----------------- | ------ | ------ | ----------- |
+| 0x00   | 2 / Big           | uint16 | size   | Size of the rest of the message including the opcode field but not including the size field.|
+| 0x02   | 2 / Little        | uint16 | opcode | Opcode that determines which fields the message contains.|
+
+### Body
+
+| Offset | Size / Endianness | Type | Name | Description | Comment |
+| ------ | ----------------- | ---- | ---- | ----------- | ------- |
+| 0x04 | - / - | [PackedGuid](../spec/packed-guid.md) | guid |  |  |
+| - | 4 / Little | u32 | counter |  |  |
