+++ wowm_language/src/docs/smsg_standstate_update.md
@@ -1,6 +1,6 @@
 # SMSG_STANDSTATE_UPDATE
 
-## Client Version 1.12
+## Client Version 1.12, Client Version 3.3.5
 
 ### Wowm Representation
